CVE: work around libcurl-native CA cert file bug
This works around the failing CVE database update because of
a failed SSL certificate check caused by the invalid file path
embedded in libcurl (https://bugzilla.yoctoproject.org/show_bug.cgi?id=9883)

isafw.py ensures that ca-certificates-native is installed and then
instructs plugins to use that ca-certificate.crt file. This is only
needed for the cve-check-tool, but gets set up for all plugins just in
case.

That path is then passed through to the patched cve-check-tool, which
overrides the default (and broken) path in libcurl.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>