isafw.bbclass: speed up build by running image analysis as task
Image analysis is in the critical path for completing a build in Ostro
OS because it runs as part of do_image_complete, and the analysis is
relatively slow (minutes).

Moving the check into a task allows image creation and the ensuing
swupd update calculation to start sooner, while the image analysis
runs in parallel.

In addition, building just the image (do_image_complete) instead of
the entire image recipe (do_build) no longer triggers the image check,
which makes the default content of the build more useful: in Ostro,
meta-swupd creates several virtual image recipes for which it runs
do_image_complete. We could disable checking via image black- or
whitelisting, but not running the task at all by default is easier.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>