ISA_cve: check cve-check-tool presence only once
The class conscrutor spawns a new shell process to check
cve-check-tool presence every time the class init is called.
These checks are redundant (and the check is known to fail).

cve-check-tool is only used in the last step when the report
is processed based on the faux file so the tool presence is
important only then.

In fact, there's no need to separately check the presences.
We'll just try to call it from the shell and log the error. In
case cve-check-tool is missing, the log file gives:
"/bin/sh: 1: cve-check-tool: not found".

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>